feat: Tabs on payment page
