+++ scouts_finances_flutter/lib/payments/matched_view.dart
@@ -1,0 +1,117 @@
+import 'package:flutter/material.dart';
+import 'package:scouts_finances_client/scouts_finances_client.dart';
+import 'package:scouts_finances_flutter/extensions/name.dart';
+import 'package:scouts_finances_flutter/extensions/payment_method.dart';
+import 'package:scouts_finances_flutter/main.dart';
+import 'package:scouts_finances_flutter/payments/shared.dart';
+import 'package:scouts_finances_flutter/payments/single_payment.dart';
+
+class MatchedView extends StatefulWidget {
+  const MatchedView({super.key});
+
+  @override
+  State<MatchedView> createState() => _MatchedViewState();
+}
+
+class _MatchedViewState extends State<MatchedView> {
+  late List<Payment> matchedPayments;
+  String? err;
+  bool loading = true;
+  String query = '';
+
+  final ScrollController _scrollController = ScrollController();
+
+  @override
+  void dispose() {
+    _scrollController.dispose();
+    super.dispose();
+  }
+
+  void _getPayments() async {
+    try {
+      final result = await client.payment.getPayments();
+
+      final matchedPayments = result.where((p) => p.parent != null).toList();
+      matchedPayments.sort((a, b) => a.date.compareTo(b.date));
+
+      setState(() {
+        this.matchedPayments = matchedPayments;
+        loading = false;
+      });
+    } catch (e) {
+      setState(() {
+        err =
+            'Failed to load payments. Are you connected to the internet? If this error persists, please contact the developers.';
+        loading = false;
+      });
+    }
+  }
+
+  @override
+  void initState() {
+    super.initState();
+    _getPayments();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    // Get which payments are action required
+    if (loading) {
+      return const Center(child: CircularProgressIndicator());
+    }
+    if (err != null) {
+      return Padding(
+          padding: const EdgeInsets.all(16.0),
+          child: Center(
+              child: Text(err!,
+                  style: const TextStyle(color: Colors.red, fontSize: 16))));
+    }
+
+    // Filter payments based on the search query
+    List<Payment> filteredMatchedPayments = matchedPayments
+        .where((payment) =>
+            payment.payee.toLowerCase().contains(query.toLowerCase()) ||
+            payment.amount.toString().contains(query) ||
+            payment.date.toLocal().toString().contains(query))
+        .toList();
+
+    List<Card> matchedPaymentCards = filteredMatchedPayments.map((payment) {
+      return toCard(context, payment);
+    }).toList();
+
+    SearchBar searchBar = SearchBar(
+      onChanged: (String value) {
+        setState(() {
+          query = value;
+        });
+      },
+      leading: Padding(
+        padding: const EdgeInsets.only(left: 8.0),
+        child: const Icon(Icons.search),
+      ),
+      hintText: 'Search by payee, amount, date...',
+    );
+
+    return SingleChildScrollView(
+        controller: _scrollController,
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            searchBar,
+            const SizedBox(height: 16.0),
+            ...matchedPaymentCards,
+            if (matchedPaymentCards.isEmpty)
+              Padding(
+                padding: const EdgeInsets.all(16.0),
+                child: Center(
+                  child: Text(
+                    'No matched payments found.',
+                    style: Theme.of(context).textTheme.bodyLarge,
+                  ),
+                ),
+              ),
+            const SizedBox(height: 128.0),
+          ],
+        ));
+  }
+}
