+++ scouts_finances_flutter/lib/payments/unmatched_view.dart
@@ -1,0 +1,114 @@
+import 'package:flutter/material.dart';
+import 'package:scouts_finances_client/scouts_finances_client.dart';
+import 'package:scouts_finances_flutter/main.dart';
+import 'package:scouts_finances_flutter/payments/shared.dart';
+
+class UnmatchedView extends StatefulWidget {
+  const UnmatchedView({super.key});
+
+  @override
+  State<UnmatchedView> createState() => _UnmatchedViewState();
+}
+
+class _UnmatchedViewState extends State<UnmatchedView> {
+  late List<Payment> unmatchedPayments;
+  String? err;
+  bool loading = true;
+  String query = '';
+
+  final ScrollController _scrollController = ScrollController();
+
+  @override
+  void dispose() {
+    _scrollController.dispose();
+    super.dispose();
+  }
+
+  void _getPayments() async {
+    try {
+      final result = await client.payment.getPayments();
+
+      final unmatchedPayments = result.where((p) => p.parent == null).toList();
+      unmatchedPayments.sort((a, b) => a.date.compareTo(b.date));
+
+      setState(() {
+        this.unmatchedPayments = unmatchedPayments;
+        loading = false;
+      });
+    } catch (e) {
+      setState(() {
+        err =
+            'Failed to load payments. Are you connected to the internet? If this error persists, please contact the developers.';
+        loading = false;
+      });
+    }
+  }
+
+  @override
+  void initState() {
+    super.initState();
+    _getPayments();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    // Get which payments are action required
+    if (loading) {
+      return const Center(child: CircularProgressIndicator());
+    }
+    if (err != null) {
+      return Padding(
+          padding: const EdgeInsets.all(16.0),
+          child: Center(
+              child: Text(err!,
+                  style: const TextStyle(color: Colors.red, fontSize: 16))));
+    }
+
+    // Filter payments based on the search query
+    List<Payment> filteredUnmatchedPayments = unmatchedPayments
+        .where((payment) =>
+            payment.payee.toLowerCase().contains(query.toLowerCase()) ||
+            (payment.amount / 100).toString().contains(query) ||
+            payment.date.toLocal().toString().contains(query))
+        .toList();
+
+    List<Card> unmatchedPaymentCards = filteredUnmatchedPayments.map((payment) {
+      return toCard(context, payment);
+    }).toList();
+
+    SearchBar searchBar = SearchBar(
+      onChanged: (String value) {
+        setState(() {
+          query = value;
+        });
+      },
+      leading: Padding(
+        padding: const EdgeInsets.only(left: 8.0),
+        child: const Icon(Icons.search),
+      ),
+      hintText: 'Search by payee, amount, date...',
+    );
+
+    return SingleChildScrollView(
+        controller: _scrollController,
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            searchBar,
+            const SizedBox(height: 16.0),
+            ...unmatchedPaymentCards,
+            if (unmatchedPaymentCards.isEmpty)
+              Padding(
+                padding: const EdgeInsets.all(16.0),
+                child: Center(
+                  child: Text(
+                    'No matched payments found.',
+                    style: Theme.of(context).textTheme.bodyLarge,
+                  ),
+                ),
+              ),
+            const SizedBox(height: 128.0),
+          ],
+        ));
+  }
+}
